Change latency representations to long as that's closr to what alsa uses.

+++ common.h
@@ -85,9 +85,9 @@ typedef struct {
   char *logfile;
   char *errfile;
   char *configfile;
-  int32_t audio_backend_buffer_desired_length; // this will be the desired number of frames in the
+  long audio_backend_buffer_desired_length; // this will be the desired number of frames in the
                                             // audio backend buffer -- the DAC buffer for ALSA
-  int32_t audio_backend_latency_offset; // this will be the offset to compensate for any fixed latency
+  long audio_backend_latency_offset; // this will be the offset to compensate for any fixed latency
   uint32_t volume_range_db; // the range, in dB, from max dB to min dB. Zero means use the mixer's native range.
                                      // there might be in the audio
 } shairport_cfg;
